hole_puncher - Some peer cleanup
