+++ peer/src/hooks/mod.rs
@@ -1,0 +1,29 @@
+mod data_for_block;
+mod explore_network;
+mod found_block;
+mod holepuncher_ack;
+mod new_block;
+mod peer_ack;
+mod peer;
+mod ping;
+mod pong;
+mod possible_block;
+mod state;
+mod sync_peers;
+mod validate_hash;
+mod validated_hash;
+
+pub use self::data_for_block::on_data_for_block;
+pub use self::explore_network::on_explore_network;
+pub use self::found_block::on_found_block;
+pub use self::holepuncher_ack::on_register_hole_puncher_ack;
+pub use self::new_block::on_new_block;
+pub use self::peer_ack::on_register_peer_ack;
+pub use self::peer::on_register_peer;
+pub use self::ping::on_ping;
+pub use self::pong::on_pong;
+pub use self::possible_block::on_possible_block;
+pub use self::state::State;
+pub use self::sync_peers::on_sync_peers;
+pub use self::validate_hash::on_validate_hash;
+pub use self::validated_hash::on_validated_hash;
