+++ peer/src/hooks/sync_peers.rs
@@ -1,0 +1,27 @@
+use blockchain_hooks::ApplicationState;
+use blockchain_protocol::BlockchainProtocol;
+use blockchain_protocol::payload::SyncPeersPayload;
+
+use hooks::State;
+
+pub fn on_sync_peers(state: ApplicationState<State>) {
+    let message = BlockchainProtocol::<SyncPeersPayload>::from_bytes(&state.payload_buffer).expect("Parsing the protocol should be successful");
+
+    {
+        let mut state_lock = state.state.lock().expect("Locking should be successful");
+
+        for new_peer in message.payload.peers {
+            let mut is_peer_known = false;
+
+            for peer in state_lock.peers.clone() {
+                if peer == new_peer {
+                    is_peer_known = true;
+                }
+            }
+
+            if !is_peer_known {
+                state_lock.peers.push(new_peer);
+            }
+        }
+    }
+}
